Move PDF upload to /audit from /report_submission
