+++ backend/report_submission/templates/report_submission/upload-page.html
@@ -54,5 +54,5 @@
     <!-- Pass variables to JS as given alias -->
     {{ report_id | json_script:'sac_id' }}
     {{ view_id | json_script:'view_id' }}
-    <script src="{% static 'compiled/js/upload-xlsx.js' %}"></script>
+    <script src="{% static 'compiled/js/upload-page.js' %}"></script>
 {% endblock content %}
